Rewrite some mapper functions, other improvements
NROM is now somewhat functional. There are several changes still to
be made to the CPU implementation, and the functionality of MMC1
is basically still unknown. However for about reasonable amount of
opcodes, the output of the emulator running a test ROM matches a
known correct value, which is promising.